microservices/indeed_data_ingestion/requirements.txt
-e git+https://github.com/velkoz-data-ingestion/velkozz_dev_api.git#egg=velkozz-developer-api
python-dotenv
requests
APScheduler
beautifulsoup4